# src/main/java/com/example/bankcards/dto/AuthRequest.java
package com.example.bankcards.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.validation.constraints.NotBlank;

public record AuthRequest(
        @Schema(description = "Имя пользователя",
                example = "user1",
                requiredMode = Schema.RequiredMode.REQUIRED)
        @NotBlank(message = "Имя пользователя не может быть пустым")
        String username,

        @Schema(description = "Пароль пользователя",
                example = "user123",
                requiredMode = Schema.RequiredMode.REQUIRED)
        @NotBlank(message = "Пароль не может быть пустым")
        String password
) {}